Set stricter compiler flags
There were some comparisons between signed and unsigned integers as
well as unused code and missing preprocessor defines. The new flags
helped finding those issues. There are still some open 'issues' marked
with 'TODO:', so simply grep using something like 'grep -r TODO: .'.
Some of the issues might be shortcomings of C, some might be fixable.

+++ src/voicetarget.c
@@ -45,7 +45,7 @@ void Voicetarget_add_session(client_t *client, int targetId, int sessionId)
 			int i;
 			vt = list_get_entry(itr, voicetarget_t, node);
 			for (i = 0; i < TARGET_MAX_SESSIONS; i++) {
-				if (vt->sessions[i] == -1) {
+				if (vt->sessions[i] == (uint32_t)-1) {
 					vt->sessions[i] = sessionId;
 					Log_debug("Adding session ID %d to voicetarget ID %d", sessionId, targetId);
 					return;
